archival/storage: extract offset to file position utils to storage
